+++ src/neo4j/_sync/_debug/_concurrency_check.py
@@ -129,7 +129,7 @@ class NonConcurrentMethodChecker:
                             tbs = deepcopy(self.__tracebacks)
                     if acquired:
                         try:
-                            item = iter_.__next__()
+                            item = next(iter_)
                         except StopIteration:
                             return
                         finally: