+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 See also https://github.com/neo4j/neo4j-python-driver/wiki for a full changelog.
 
 ## NEXT RELEASE
+- Python 3.7, 3.8, and 3.9 support has been dropped.
 - Changed errors raised under certain circumstances
   - `ConfigurationError` if the passed `auth` parameters is not valid (instead of `AuthError`)
     - This improves the differentiation between `DriverError` for client-side errors and `Neo4jError` for server-side errors.